# modules/academy-api/src/main/resources/openapi.yaml
openapi: 3.0.1
info:
  title: Academy Management System API
  version: 1.0.0
  description: |
    REST API for Academy Management System. This API provides endpoints for managing students, batches, classes, mentors, and mentor sessions.
    
    The API supports:
    - Student registration and management
    - Batch creation and management
    - Class scheduling and management
    - Mentor management
    - Mentor session booking and management
    
    All create operations publish Kafka events for event-driven architecture.
  contact:
    name: Academy Development Team
    email: support@academy.com
servers:
  - url: http://localhost:8080
    description: Local development server
  - url: https://academy-backend-dev.example.com
    description: Development environment
  - url: https://academy-backend.example.com
    description: Production environment
security:
  - bearerAuth: []
...
paths:
...
  /api/students:
    get:
      summary: Get all students
      operationId: getAllStudents
      tags:
        - Students
      security:
        - bearerAuth: []
      parameters:
        - name: batchId
          in: query
          required: false
          schema:
            type: integer
            format: int64
          description: Filter students by batch ID
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Student'
        '500':
          $ref: '#/components/responses/InternalServerError'
    
    post:
      summary: Create a new student
      operationId: createStudent
      tags:
        - Students
      description: Register a new student and publish student.registered Kafka event
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StudentInput'
      responses:
        '201':
          description: Student created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Student'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/students/paged:
    get:
      summary: Get all students (paginated)
      operationId: getAllStudentsPaged
      tags:
        - Students
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            format: int32
            default: 0
          description: Page number (0-indexed)
        - name: size
          in: query
          required: false
          schema:
            type: integer
            format: int32
            default: 20
          description: Page size
        - name: sort
          in: query
          required: false
          schema:
            type: string
          description: Sort criteria (e.g., "name,asc")
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageStudent'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/students/{id}:
    get:
      summary: Get student by ID
      operationId: getStudentById
      tags:
        - Students
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/StudentId'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Student'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    
    put:
      summary: Update student
      operationId: updateStudent
      tags:
        - Students
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/StudentId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StudentInput'
      responses:
        '200':
          description: Student updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Student'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    
    delete:
      summary: Delete student
      operationId: deleteStudent
      tags:
        - Students
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/StudentId'
      responses:
        '204':
          description: Student deleted successfully
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /api/mentors:
    get:
      summary: Get all mentors
      operationId: getAllMentors
      tags:
        - Mentors
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Mentor'
        '500':
          $ref: '#/components/responses/InternalServerError'
    
    post:
      summary: Create a new mentor
      operationId: createMentor
      tags:
        - Mentors
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MentorInput'
      responses:
        '201':
          description: Mentor created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Mentor'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/mentors/{id}:
    get:
      summary: Get mentor by ID
      operationId: getMentorById
      tags:
        - Mentors
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/MentorId'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Mentor'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    
    put:
      summary: Update mentor
      operationId: updateMentor
      tags:
        - Mentors
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/MentorId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MentorInput'
      responses:
        '200':
          description: Mentor updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Mentor'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    
    delete:
      summary: Delete mentor
      operationId: deleteMentor
      tags:
        - Mentors
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/MentorId'
      responses:
        '204':
          description: Mentor deleted successfully
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/mentor-sessions:
    get:
      summary: Get all mentor sessions
      operationId: getAllMentorSessions
      tags:
        - Mentor Sessions
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MentorSession'
        '500':
          $ref: '#/components/responses/InternalServerError'
    
    post:
      summary: Create a new mentor session
      operationId: createMentorSession
      tags:
        - Mentor Sessions
      description: Book a mentor session and publish mentor.session.created Kafka event
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MentorSessionInput'
      responses:
        '201':
          description: Mentor session created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MentorSession'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/mentor-sessions/{id}:
    get:
      summary: Get mentor session by ID
      operationId: getMentorSessionById
      tags:
        - Mentor Sessions
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/MentorSessionId'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MentorSession'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    
    put:
      summary: Update mentor session
      operationId: updateMentorSession
      tags:
        - Mentor Sessions
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/MentorSessionId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MentorSessionInput'
      responses:
        '200':
          description: Mentor session updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MentorSession'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    
    delete:
      summary: Delete mentor session
      operationId: deleteMentorSession
      tags:
        - Mentor Sessions
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/MentorSessionId'
      responses:
        '204':
          description: Mentor session deleted successfully
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT token authentication. 
        Token should be sent in the Authorization header with "Bearer " prefix.
        Header name is configurable via jwt.header.name (default: Authorization).
        Header prefix is configurable via jwt.header.prefix (default: Bearer ).
        
        Example:
        Authorization: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
  parameters:
    StudentId:
      name: id
      in: path
      required: true
      schema:
        type: integer
        format: int64
      description: Student ID
...
    MentorId:
      name: id
      in: path
      required: true
      schema:
        type: integer
        format: int64
      description: Mentor ID
    MentorSessionId:
      name: id
      in: path
      required: true
      schema:
        type: integer
        format: int64
      description: Mentor session ID
...
  schemas:
    Student:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: Student ID
        name:
          type: string
          description: Student full name
          example: John Doe
        email:
          type: string
          format: email
          description: Student email address
          example: john.doe@example.com
        graduationYear:
          type: integer
          format: int32
          description: Year of graduation
          example: 2024
        universityName:
          type: string
          description: Name of the university
          example: State University
        phoneNumber:
          type: string
          description: Contact phone number
          example: "+1-555-123-4567"
        batchId:
          type: integer
          format: int64
          description: ID of the batch the student belongs to
        buddyId:
          type: integer
          format: int64
          description: ID of the student's buddy
    StudentInput:
      type: object
      required:
        - name
        - email
      properties:
        name:
          type: string
          description: Student full name
          example: John Doe
        email:
          type: string
          format: email
          description: Student email address
          example: john.doe@example.com
        graduationYear:
          type: integer
          format: int32
          description: Year of graduation
          example: 2024
        universityName:
          type: string
          description: Name of the university
          example: State University
        phoneNumber:
          type: string
          description: Contact phone number
          example: "+1-555-123-4567"
        batchId:
          type: integer
          format: int64
          description: ID of the batch the student belongs to
        buddyId:
          type: integer
          format: int64
          description: ID of the student's buddy
...
    Mentor:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: Mentor ID
        name:
          type: string
          description: Mentor full name
          example: Alice Johnson
        currentCompany:
          type: string
          description: Current company where mentor works
          example: Tech Corp
    MentorInput:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          description: Mentor full name
          example: Alice Johnson
        currentCompany:
          type: string
          description: Current company where mentor works
          example: Tech Corp
    MentorSession:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: Mentor session ID
        time:
          type: string
          format: date-time
          description: Session scheduled time
          example: "2024-03-20T15:00:00Z"
        durationMinutes:
          type: integer
          format: int32
          description: Session duration in minutes
          example: 60
        studentId:
          type: integer
          format: int64
          description: Student ID
        mentorId:
          type: integer
          format: int64
          description: Mentor ID
        studentRating:
          type: integer
          format: int32
          description: Student's rating of the session (1-5)
          minimum: 1
          maximum: 5
        mentorRating:
          type: integer
          format: int32
          description: Mentor's rating of the session (1-5)
          minimum: 1
          maximum: 5
    MentorSessionInput:
      type: object
      required:
        - time
        - durationMinutes
        - studentId
        - mentorId
      properties:
        time:
          type: string
          format: date-time
          description: Session scheduled time
          example: "2024-03-20T15:00:00Z"
        durationMinutes:
          type: integer
          format: int32
          description: Session duration in minutes
          example: 60
        studentId:
          type: integer
          format: int64
          description: Student ID
        mentorId:
          type: integer
          format: int64
          description: Mentor ID
        studentRating:
          type: integer
          format: int32
          description: Student's rating of the session (1-5)
          minimum: 1
          maximum: 5
        mentorRating:
          type: integer
          format: int32
          description: Mentor's rating of the session (1-5)
          minimum: 1
          maximum: 5
    PageStudent:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/Student'
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
        number:
          type: integer
          format: int32
        first:
          type: boolean
        last:
          type: boolean
...
    ProblemDetail:
      type: object
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
          format: int32
        detail:
          type: string
        instance:
          type: string
          format: uri
  responses:
    BadRequest:
      description: Bad Request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
    InternalServerError:
      description: Internal Server Error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'